more work on chart component docs
